Format the label/comboBox with slider and spinner into a function so later algorithms can apply it. Change the type of radius from Int to Double to generic the function above.

+++ src/main/kotlin/controller/EngineController.kt
@@ -41,7 +41,7 @@ class EngineController : Controller() {
 
     fun flipVertical() = engine.transform(FlipVertical())
 
-    fun blur(radius: Int, type: BlurType) = engine.adjust(type.name, radius.toDouble())
+    fun blur(radius: Double, type: BlurType) = engine.adjust(type.name, radius)
 
     fun sharpen() = engine.transform(Sharpen())
 }
